requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.3
django-adminlte3==0.1.6
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.54.4
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3

